Multiplayer Physics Update
Attempt to improve playability online by only processing physics on the player on the authority, which should be the server. This seems to work well based on short testing.

Without this, the player rubber bands around badly.
With this, other players may get behind on what a remote player is doing, but the local player's play will fee smooth and normal.