New method to gracefully restart the server.
Godot headless always terminates instantly upon any signal in Linux.
Instead, a client can now ask the server to shut down.
The client must include a password which the server will randomly generate and save to tis own config file.
The "client" that asks the server to shut down should be run from the command line on the same host as the server, which will allow it to pick up the password from the same file.